merge test

+++ ShopFloorPlacementPlanner/Staff.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Data.SqlClient;
 
 namespace ShopFloorPlacementPlanner
@@ -26,13 +22,7 @@ namespace ShopFloorPlacementPlanner
             SqlCommand cmd = new SqlCommand("SELECT id from dbo.view_full_name where fullname = @fullname", conn);
             cmd.Parameters.AddWithValue("@fullname", _fullname);
 
-            _staffID = Convert.ToInt16(cmd.ExecuteScalar().ToString());
-
+            _staffID = Convert.ToInt32(cmd.ExecuteScalar().ToString());
         }
-
-
-     
-
-
     }
 }
